ログイン機能の開発

+++ todo_laravel/app/Http/Controllers/TaskController.php
@@ -3,10 +3,21 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+
 use App\Models\Task;
 
 class TaskController extends Controller
 {
+    /**
+     * コンストラクタ
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     /**
      * タスク一覧
      *
@@ -15,7 +26,8 @@ class TaskController extends Controller
      */
     public function index(Request $request)
     {
-        $tasks = Task::orderBy('created_at', 'asc')->get();
+        //$tasks = Task::orderBy('created_at', 'asc')->get();
+        $tasks = $request->user()->tasks()->get();
         return view('tasks.index', [
             'tasks' => $tasks,
         ]);
@@ -34,9 +46,8 @@ class TaskController extends Controller
         ]);
 
         // タスク作成
-        Task::create([
-            'user_id' => 0,
-            'name' => $request->name
+        $request->user()->tasks()->create([
+            'name' => $request->name,
         ]);
 
         return redirect('/tasks');
@@ -45,13 +56,17 @@ class TaskController extends Controller
     /**
      * タスク削除
      *
-     * @param int $taskId
+     * @param Request $request
+     * @param Task $task
      * @return Response
      */
     public function delete($taskId)
     {
         $task = Task::findOrFail($taskId);
+        $this->authorize('delete', $task);
         $task->delete();
+
+
         return redirect('/tasks');
     }
 }
